requirements.txt
# --- Core ---
oandapyV20==0.7.2
openai==1.78.0           # GPT-4o / GPT-4o-mini
pandas==2.2.3
numpy==1.24.4            # 3.11 compatible wheel
websockets==12.0
apscheduler==3.10.4
PyYAML==6.0.1
pydantic==2.7.1
toml
annotated-types==0.7.0
anyio==4.9.0
urllib3==2.4.0
setuptools>=70.0.0

# --- Google Cloud SDKs ---
google-cloud-storage==2.15.0
google-cloud-pubsub==2.21.0
google-cloud-secret-manager==2.18.3
google-cloud-logging==3.10.0
google-cloud-bigquery==3.19.0
google-cloud-firestore==2.16.0

# --- HTTP / RSS ---
requests==2.32.3
feedparser==6.0.11        # RSS parsing
httpx==0.27.0

# --- Dev / Test ---
pytest==8.3.2
pytest-asyncio==0.23.8
flake8==7.0.0
black==24.4.2
ruff==0.4.4
pre-commit==3.7.1

# --- Web ---
Flask==3.0.3
